Added most commonly used files in this project
TX1_OFDM_file_BASELINE.grc is a working OFDM modulation scheme for sending and receiving files. Things that might need to be changed are the file location in the file sink block, and the USRP sinks' IP addresses (if using multiple SDRs). 
PER.py is a packet error rate calculator script.